MAINT: Renamed solver_mixin to porepy_integration
