+++ src/pp_solvers/__init__.py
@@ -5,12 +5,12 @@ from .mat_utils import *
 from .petsc_utils import *
 from .plot_linear_system import *
 from .preconditioners import *
-from .solver_mixin import *
+from .porepy_integration import *
 
 __all__ = []
 __all__.extend(block_linear_system.__all__)
 __all__.extend(mat_utils.__all__)
-__all__.extend(solver_mixin.__all__)
+__all__.extend(porepy_integration.__all__)
 __all__.extend(preconditioners.__all__)
 __all__.extend(petsc_utils.__all__)
 __all__.extend(plot_linear_system.__all__)
